move about post defs to app directories blog.py
